Added Layer merging, partialized document and added layer rename shortcut
